backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-genai==0.7.0
neo4j==5.28.3
psycopg2-binary==2.9.11
pgvector==0.2.5
pdfplumber==0.11.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.0.0
httpx==0.28.1
slowapi==0.1.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
